user policy finished

+++ app/Http/Controllers/UserController.php
@@ -7,6 +7,7 @@ use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -84,6 +85,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        abort_unless(Gate::allows('update', $user), 403);
+
         return view('admin.editUser')->with([
             'user' => $user,
             'message_success' => Session::get('message_success'),
@@ -102,6 +105,8 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        abort_unless(Gate::allows('update', $user), 403);
+
            $request->validate([
             'name' => 'required|min:2',
             'email' => 'required|min:5',
@@ -136,6 +141,8 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        abort_unless(Gate::allows('delete', $user), 403);
+
            $oldName = $user->name;
         $user->delete();
         return $this->index()->with(
